refactor of search-specific logic into PCP app
