+++ plan/views.py
@@ -1,3 +1,13 @@
 from django.shortcuts import render
 
-# Create your views here.
+from courses.views import CourseList
+from .search import TypedSearchBackend
+
+
+class CourseListSearch(CourseList):
+    filter_backends = (TypedSearchBackend, )
+    search_fields = ('full_code', 'title', 'sections__instructors__name')
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        return queryset
